adding movie stats without title

+++ src/main/scala/org/esgi/project/streaming/models/MovieStat.scala
@@ -3,9 +3,9 @@ package org.esgi.project.streaming.models
 import play.api.libs.json.{Json, OFormat}
 
 case class MovieStat(
-                      start_only: Int,
-                      half: Int,
-                      full: Int
+                      start_only: Long,
+                      half: Long,
+                      full: Long
                     )
